Refactor Link Utils classes
More clear distinguisging between links helper and intents helper.
