+++ app/src/main/java/com/github/fo2rist/mclaren/ui/WebPreviewFragment.java
@@ -16,7 +16,7 @@ import android.webkit.WebViewClient;
 import com.github.fo2rist.mclaren.R;
 import com.github.fo2rist.mclaren.ui.utils.McLarenHtmlUtils;
 
-import static com.github.fo2rist.mclaren.ui.utils.FeedLinkUtils.MCLAREN_COM;
+import static com.github.fo2rist.mclaren.ui.utils.LinkUtils.MCLAREN_COM;
 
 
 public class WebPreviewFragment extends Fragment {
