+++ app/src/test/java/com/github/fo2rist/mclaren/ui/utils/LinkUtilsTagsTest.java
@@ -12,7 +12,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.runners.Parameterized.*;
 
 @RunWith(Parameterized.class)
-public class FeedLinkUtilsTagsTest {
+public class LinkUtilsTagsTest {
 
     @Parameters(name = "Test tag resolved \"{0}\" -> \"{1}\"")
     public static List<String[]> data() {
@@ -32,28 +32,28 @@ public class FeedLinkUtilsTagsTest {
 
     @Test
     public void testTwitterMentionResolved() throws Exception {
-        String link = FeedLinkUtils.getFeedMentionLink(FeedItems.TWITTER_GALLERY_ITEM, sourceTag);
+        String link = LinkUtils.getFeedMentionLink(FeedItems.TWITTER_GALLERY_ITEM, sourceTag);
 
         assertEquals("https://twitter.com/" + targetTag, link);
     }
 
     @Test
     public void testInstagramMentionResolved() throws Exception {
-        String link = FeedLinkUtils.getFeedMentionLink(FeedItems.INSTAGRAM_GALLERY_ITEM, sourceTag);
+        String link = LinkUtils.getFeedMentionLink(FeedItems.INSTAGRAM_GALLERY_ITEM, sourceTag);
 
         assertEquals("https://www.instagram.com/" + targetTag, link);
     }
 
     @Test
     public void testTwitterHashTagResolved() throws Exception {
-        String link = FeedLinkUtils.getFeedHashtagLink(FeedItems.TWITTER_GALLERY_ITEM, sourceTag);
+        String link = LinkUtils.getFeedHashtagLink(FeedItems.TWITTER_GALLERY_ITEM, sourceTag);
 
         assertEquals("https://twitter.com/hashtag/" + targetTag, link);
     }
 
     @Test
     public void testInstagramHashTagResolved() throws Exception {
-        String link = FeedLinkUtils.getFeedHashtagLink(FeedItems.INSTAGRAM_GALLERY_ITEM, sourceTag);
+        String link = LinkUtils.getFeedHashtagLink(FeedItems.INSTAGRAM_GALLERY_ITEM, sourceTag);
 
         assertEquals("https://www.instagram.com/explore/tags/" + targetTag, link);
     }
