+++ app/src/test/java/com/github/fo2rist/mclaren/ui/utils/LinkUtilsUrlsTest.java
@@ -8,11 +8,11 @@ import org.junit.runners.JUnit4;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(JUnit4.class)
-public class FeedLinkUtilsUrlsTest {
+public class LinkUtilsUrlsTest {
 
     @Test
     public void testUnknonwSourceTreatAsMclarenF1Link() throws Exception {
-        String link = FeedLinkUtils.getFeedMentionLink(FeedItems.MCLAREN_ARTICLE_ITEM,
+        String link = LinkUtils.getFeedMentionLink(FeedItems.MCLAREN_ARTICLE_ITEM,
                 "any");
 
         assertEquals("http://www.mclaren.com/formula1/", link);
